+++ DigitalLearningSolutions.Data.Tests/DataServices/DiagnosticAssessmentDataServiceTests/DiagnosticAssessmentDataServiceOtherItemAndSectionTests.cs
@@ -1,4 +1,4 @@
-﻿namespace DigitalLearningSolutions.Data.Tests.Services.DiagnosticAssessmentDataServiceTests
+﻿namespace DigitalLearningSolutions.Data.Tests.DataServices.DiagnosticAssessmentDataServiceTests
 {
     using System.Linq;
     using System.Transactions;
