+++ DigitalLearningSolutions.Data.Tests/DataServices/DiagnosticAssessmentDataServiceTests/DiagnosticAssessmentDataServiceTests.cs
@@ -1,11 +1,10 @@
-﻿namespace DigitalLearningSolutions.Data.Tests.Services.DiagnosticAssessmentDataServiceTests
+﻿namespace DigitalLearningSolutions.Data.Tests.DataServices.DiagnosticAssessmentDataServiceTests
 {
     using System;
     using System.Linq;
     using System.Transactions;
+    using DigitalLearningSolutions.Data.DataServices;
     using DigitalLearningSolutions.Data.Models.DiagnosticAssessment;
-    using DigitalLearningSolutions.Data.Services;
-    using DigitalLearningSolutions.Data.Tests.Helpers;
     using DigitalLearningSolutions.Data.Tests.TestHelpers;
     using FakeItEasy;
     using FluentAssertions;
