+++ DigitalLearningSolutions.Data/DataServices/CourseContentService.cs
@@ -1,4 +1,4 @@
-﻿namespace DigitalLearningSolutions.Data.Services
+﻿namespace DigitalLearningSolutions.Data.DataServices
 {
     using System;
     using System.Data;
