+++ DigitalLearningSolutions.Data/DataServices/DiagnosticAssessmentDataService.cs
@@ -1,57 +1,57 @@
-﻿namespace DigitalLearningSolutions.Data.Services
+﻿namespace DigitalLearningSolutions.Data.DataServices
 {
+    using System;
     using System.Data;
     using System.Linq;
     using Dapper;
-    using DigitalLearningSolutions.Data.Models.PostLearningAssessment;
+    using DigitalLearningSolutions.Data.Models.DiagnosticAssessment;
     using Microsoft.Extensions.Logging;
 
-    public interface IPostLearningAssessmentService
+    public interface IDiagnosticAssessmentDataService
     {
-        PostLearningAssessment? GetPostLearningAssessment(int customisationId, int candidateId, int sectionId);
-        PostLearningContent? GetPostLearningContent(int customisationId, int sectionId);
+        DiagnosticAssessment? GetDiagnosticAssessment(int customisationId, int candidateId, int sectionId);
+        DiagnosticContent? GetDiagnosticContent(int customisationId, int sectionId);
     }
 
-    public class PostLearningAssessmentService : IPostLearningAssessmentService
+    public class DiagnosticAssessmentDataService : IDiagnosticAssessmentDataService
     {
         private readonly IDbConnection connection;
-        private readonly ILogger<PostLearningAssessmentService> logger;
+        private readonly ILogger<DiagnosticAssessmentDataService> logger;
 
-        public PostLearningAssessmentService(IDbConnection connection, ILogger<PostLearningAssessmentService> logger)
+        public DiagnosticAssessmentDataService(IDbConnection connection, ILogger<DiagnosticAssessmentDataService> logger)
         {
             this.connection = connection;
             this.logger = logger;
         }
 
-        public PostLearningAssessment? GetPostLearningAssessment(int customisationId, int candidateId, int sectionId)
+        public DiagnosticAssessment? GetDiagnosticAssessment(int customisationId, int candidateId, int sectionId)
         {
-            // NextSectionID is the ID of the next section in the course, according to SectionNumber
-            // or null if the last in the course.
+            // NextTutorialID is the ID of the first tutorial in the section, according to Tutorials.OrderBy
+            // or null if the last in the section. Similar for NextSectionID, using SectionID and SectionNumber
+
+            // Find these by making a list of other tutorials in the course, to then find tutorials in this section,
+            // and other sections (because a section must contain at least one tutorial), using a similar approach to
+            // the one used in the TutorialContentDataService.
 
             // Using this list of other tutorials in the course we can work out if there is another item in the
             // section (if there is an viewable tutorial, or a post learning assessment, or consolidation material),
-            // and if there are other sections (valid tutorials with a different tutorial ID, or with assessments or
-            // consolidation material. See the SectionContentDataService for the definition of a valid section.
+            // and if there are other sections (valid tutorials with a different section ID, or with assessments or
+            // consolidation material). See the SectionContentDataService for the definition of a valid section.
 
-            return connection.QueryFirstOrDefault<PostLearningAssessment>(
+            DiagnosticAssessment? diagnosticAssessment = null;
+            return connection.Query<DiagnosticAssessment, DiagnosticTutorial, DiagnosticAssessment>(
                 @"  WITH CourseTutorials AS (
                     SELECT Tutorials.TutorialID,
                            Tutorials.OrderByNumber,
                            CustomisationTutorials.Status,
                            Sections.SectionID,
-                           Sections.SectionNumber,
-                           CAST (CASE WHEN CustomisationTutorials.DiagStatus = 1 AND Sections.DiagAssessPath IS NOT NULL
-                                           THEN 1
-                                      ELSE 0
-                                 END AS BIT) AS HasDiagnostic
+                           Sections.SectionNumber
                       FROM Tutorials
                            INNER JOIN Customisations
                                ON Customisations.CustomisationID = @customisationId
-
                            INNER JOIN Sections
                                ON Tutorials.SectionID = Sections.SectionID
                                AND Sections.ArchivedDate IS NULL
-
                            INNER JOIN CustomisationTutorials
                                ON CustomisationTutorials.CustomisationID = @customisationId
                                AND CustomisationTutorials.TutorialID = Tutorials.TutorialID
@@ -69,96 +69,140 @@
                       FROM Sections AS CurrentSection
                            INNER JOIN CourseTutorials
                                ON CourseTutorials.SectionID <> CurrentSection.SectionID
-
                      WHERE CurrentSection.SectionID = @sectionId
                            AND CurrentSection.SectionNumber <= CourseTutorials.SectionNumber
                            AND (
                                 CurrentSection.SectionNumber < CourseTutorials.SectionNumber
                                 OR CurrentSection.SectionID < CourseTutorials.SectionID
                            )
-
                      ORDER BY CourseTutorials.SectionNumber, CourseTutorials.SectionID
+                    ),
+                    NextTutorial AS (
+                    SELECT TOP 1
+                           CourseTutorials.SectionID AS CurrentSectionID,
+                           CourseTutorials.TutorialID AS NextTutorialID
+                      FROM CourseTutorials
+                           INNER JOIN CustomisationTutorials AS OtherCustomisationTutorials
+                              ON OtherCustomisationTutorials.CustomisationID = @customisationId
+                              AND OtherCustomisationTutorials.Status = 1
+                     WHERE CourseTutorials.SectionID = @sectionId
+                     ORDER BY CourseTutorials.OrderByNumber, CourseTutorials.TutorialID
                     )
                     SELECT
                         Applications.ApplicationName,
                         Applications.ApplicationInfo,
                         Customisations.CustomisationName,
                         Sections.SectionName,
-                        COALESCE (Attempts.BestScore, 0) AS BestScore,
-                        COALESCE (Attempts.AttemptsPL, 0) AS AttemptsPL,
-                        COALESCE (Attempts.PLPasses, 0) AS PLPasses,
-                        CAST (COALESCE (Progress.PLLocked, 0) AS bit) AS PLLocked,
-                        Applications.IncludeCertification,
+                        COALESCE (aspProgress.DiagAttempts, 0) AS DiagAttempts,
+                        COALESCE (aspProgress.DiagLast, 0) AS DiagLast,
+                        Tutorials.DiagAssessOutOf,
+                        Sections.DiagAssessPath,
+                        Customisations.DiagObjSelect,
+                        Sections.PLAssessPath,
                         Customisations.IsAssessed,
+                        Applications.IncludeCertification,
                         Progress.Completed,
                         Applications.AssessAttempts AS MaxPostLearningAssessmentAttempts,
                         Applications.PLAPassThreshold AS PostLearningAssessmentPassThreshold,
                         Customisations.DiagCompletionThreshold AS DiagnosticAssessmentCompletionThreshold,
                         Customisations.TutCompletionThreshold AS TutorialsCompletionThreshold,
-                        NextSection.NextSectionID,
+                        NextTutorial.NextTutorialId,
+                        NextSection.NextSectionId,
                         CAST (CASE WHEN EXISTS(SELECT 1
                                                  FROM CourseTutorials
                                                 WHERE SectionID <> @sectionId)
                                         THEN 1
                                    ELSE 0
                               END AS BIT) AS OtherSectionsExist,
-                        CAST (CASE WHEN Sections.ConsolidationPath IS NOT NULL
+                        CAST (CASE WHEN (Customisations.IsAssessed = 1 AND Sections.PLAssessPath IS NOT NULL)
+                                        OR Sections.ConsolidationPath IS NOT NULL
                                         THEN 1
                                    WHEN EXISTS(SELECT 1
                                                  FROM CourseTutorials
                                                 WHERE SectionID = @sectionId
-                                                      AND (Status = 1 OR HasDiagnostic = 1)
+                                                      AND Status = 1
                                               )
                                         THEN 1
                                    ELSE 0
                               END AS BIT) AS OtherItemsInSectionExist,
-                        Customisations.Password,
-                        Progress.PasswordSubmitted
+                         Customisations.Password,
+                         Progress.PasswordSubmitted,
+                        Tutorials.TutorialName,
+                        CASE WHEN Tutorials.OriginalTutorialID > 0
+                             THEN Tutorials.OriginalTutorialID
+                             ELSE Tutorials.TutorialID
+                        END AS id
                     FROM Sections
                         INNER JOIN Customisations
                             ON Customisations.ApplicationID = Sections.ApplicationID
+                            AND Customisations.Active = 1
                         INNER JOIN Applications
                             ON Applications.ApplicationID = Sections.ApplicationID
-                        LEFT JOIN NextSection
-                            ON Sections.SectionID = NextSection.CurrentSectionID
+                        INNER JOIN CustomisationTutorials
+                            ON CustomisationTutorials.CustomisationID = Customisations.CustomisationID
+                        INNER JOIN Tutorials
+                            ON CustomisationTutorials.TutorialID = Tutorials.TutorialID
+                            AND Tutorials.SectionID = Sections.SectionID
                         LEFT JOIN Progress
                             ON Progress.CustomisationID = Customisations.CustomisationID
                             AND Progress.CandidateID = @candidateId
                             AND Progress.RemovedDate IS NULL
                             AND Progress.SystemRefreshed = 0
-                        LEFT JOIN (
-                            SELECT
-                                COUNT(AssessAttemptID) AS AttemptsPL,
-                                AssessAttempts.ProgressID,
-                                AssessAttempts.SectionNumber,
-                                MAX(AssessAttempts.Score) AS BestScore,
-                                SUM(CAST(AssessAttempts.Status AS Integer)) AS PLPasses
-                            FROM AssessAttempts
-                                GROUP BY
-                                    AssessAttempts.ProgressID,
-                                    AssessAttempts.SectionNumber
-                            ) AS Attempts ON (Progress.ProgressID = Attempts.ProgressID) AND (Attempts.SectionNumber = Sections.SectionNumber)
+                        LEFT JOIN aspProgress
+                            ON aspProgress.TutorialID = CustomisationTutorials.TutorialID
+                            AND aspProgress.ProgressID = Progress.ProgressID
+                        LEFT JOIN NextSection
+                            ON Sections.SectionID = NextSection.CurrentSectionID
+                        LEFT JOIN NextTutorial
+                            ON Sections.SectionID = NextTutorial.CurrentSectionID
                     WHERE
                         Customisations.CustomisationID = @customisationId
-                        AND Customisations.IsAssessed = 1
                         AND Sections.SectionID = @sectionId
                         AND Sections.ArchivedDate IS NULL
-                        AND Sections.PLAssessPath IS NOT NULL
-                        AND Applications.DefaultContentTypeID <> 4;",
-                new { customisationId, candidateId, sectionId }
-            );
+                        AND Sections.DiagAssessPath IS NOT NULL
+                        AND CustomisationTutorials.DiagStatus = 1
+                        AND Tutorials.DiagAssessOutOf > 0
+                        AND Tutorials.ArchivedDate IS NULL
+                        AND Applications.DefaultContentTypeID <> 4
+                    ORDER BY
+                        Tutorials.OrderByNumber,
+                        id;",
+                (diagnostic, tutorial) =>
+                {
+                    if (diagnosticAssessment == null)
+                    {
+                        diagnosticAssessment = diagnostic;
+                    }
+                    else
+                    {
+                        diagnosticAssessment.DiagnosticAttempts = Math.Max(
+                            diagnosticAssessment.DiagnosticAttempts,
+                            (int)diagnostic.DiagnosticAttempts
+                        );
+                        diagnosticAssessment.SectionScore += diagnostic.SectionScore;
+                        diagnosticAssessment.MaxSectionScore += diagnostic.MaxSectionScore;
+                    }
+
+                    diagnosticAssessment.Tutorials.Add(tutorial);
+
+                    return diagnosticAssessment;
+                },
+                new { customisationId, candidateId, sectionId },
+                splitOn: "TutorialName"
+            ).FirstOrDefault();
         }
 
-        public PostLearningContent? GetPostLearningContent(int customisationId, int sectionId)
+        public DiagnosticContent? GetDiagnosticContent(int customisationId, int sectionId)
         {
-            PostLearningContent? postLearningContent = null;
-            return connection.Query<PostLearningContent, int, PostLearningContent>(
+            DiagnosticContent? diagnosticContent = null;
+            return connection.Query<DiagnosticContent, int, DiagnosticContent>(
                 @"
                     SELECT
                         Applications.ApplicationName,
                         Customisations.CustomisationName,
                         Sections.SectionName,
-                        Sections.PLAssessPath,
+                        Sections.DiagAssessPath,
+                        Customisations.DiagObjSelect,
                         Applications.PLAPassThreshold,
                         Customisations.CurrentVersion,
                         CASE WHEN Tutorials.OriginalTutorialID > 0
@@ -169,31 +213,32 @@
                         INNER JOIN Customisations
                             ON Customisations.ApplicationID = Sections.ApplicationID
                             AND Customisations.Active = 1
-                            AND Customisations.IsAssessed = 1
                         INNER JOIN Applications
                             ON Applications.ApplicationID = Sections.ApplicationID
                         INNER JOIN CustomisationTutorials
                             ON CustomisationTutorials.CustomisationID = Customisations.CustomisationID
+                            AND CustomisationTutorials.DiagStatus = 1
                         INNER JOIN Tutorials
                             ON Tutorials.TutorialID = CustomisationTutorials.TutorialID
                             AND Tutorials.SectionID = Sections.SectionID
+                            AND Tutorials.DiagAssessOutOf > 0
                             AND Tutorials.ArchivedDate IS NULL
                     WHERE
                         Customisations.CustomisationID = @customisationId
                         AND Sections.SectionID = @sectionId
                         AND Sections.ArchivedDate IS NULL
-                        AND Sections.PLAssessPath IS NOT NULL
+                        AND Sections.DiagAssessPath IS NOT NULL
                         AND Applications.DefaultContentTypeID <> 4
                     ORDER BY
                         Tutorials.OrderByNumber,
                         Tutorials.TutorialID",
-                (postLearning, tutorialId) =>
+                (diagnostic, tutorialId) =>
                 {
-                    postLearningContent ??= postLearning;
+                    diagnosticContent ??= diagnostic;
 
-                    postLearningContent.Tutorials.Add(tutorialId);
+                    diagnosticContent.Tutorials.Add(tutorialId);
 
-                    return postLearningContent;
+                    return diagnosticContent;
                 },
                 new { customisationId, sectionId },
                 splitOn: "id"
