+++ DigitalLearningSolutions.Web/ViewComponents/LogoViewComponent.cs
@@ -1,6 +1,7 @@
 namespace DigitalLearningSolutions.Web.ViewComponents
 {
     using System.Security.Claims;
+    using DigitalLearningSolutions.Data.DataServices;
     using DigitalLearningSolutions.Data.Services;
     using DigitalLearningSolutions.Web.Helpers;
     using Microsoft.AspNetCore.Mvc;
